third_party/blink/web_tests/svg/dynamic-updates/SVGFEDisplacementMapElement-dom-yChannelSelector-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'yChannelSelector' attribute of the SVGFEDisplacementMapElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS displacementMap.getAttribute('yChannelSelector') is "B"
PASS displacementMap.getAttribute('yChannelSelector') is "G"
PASS successfullyParsed is true

TEST COMPLETE
